+++ fast_mail_parser/__init__.py
@@ -1,8 +1,15 @@
-from .fast_mail_parser import ParseError, PyAttachment, PyMail, parse_email
+from .fast_mail_parser import (
+    ParseError,
+    PyAddress,
+    PyAttachment,
+    PyMail,
+    parse_email,
+)
 
 __all__ = [
     "parse_email",
     "ParseError",
     "PyMail",
     "PyAttachment",
+    "PyAddress",
 ]